+++ src/Modules/AppModule.cpp
@@ -18,51 +18,45 @@ const char *Named::getName() { return name; }
 
 size_t Named::print(Print *p, char ch) { return p->print(ch); };
 
-size_t Named::print(Print *p, char *str) { return p->print(str); };
-
-size_t Named::print(Print *p, const char *str, size_t size) {
-    char buf[size + 1];
-    strncpy(buf, str, size);
-    return p->print(buf);
-};
-
-size_t Named::print_name(Print *p) {
-    size_t n = print(p, '[');
-    n += print(p, name);
-    n += print(p, ']');
-    n += print(p, ' ');
-    return n;
+size_t Named::print(Print *p, const char *str) {
+    return print(p, (char *)str, strlen(str));
 }
 
-size_t Named::say(String &str) {
-    size_t n = print_name(out);
-    n += print(out, str.c_str(), str.length());
-    return n += out->println();
+size_t Named::print(Print *p, char *str, size_t size) {
+    size_t n = 0;
+    for (n = 0; n < size; ++n) {
+        n += print(p, str[n]);
+        if (str[n] == '\x00')
+            break;
+    }
+    return n;
 };
 
 size_t Named::say(char *str) {
-    size_t n = print_name(out);
-    n += out->print(str);
-    return n += out->println();
+    size_t n = print_ident(out, name);
+    n += print(out, str);
+    return n += print_ln(out);
 }
 
-size_t Named::sayP(PGM_P strP) {
+size_t Named::say_strP(PGM_P strP) {
+    size_t n = print_ident(out, name);
+    n += print(out, ' ');
     char buf[64];
     strcpy_P(buf, strP);
+    n += print(out, buf);
     return say(buf);
 }
 
-size_t Named::sayP_value(PGM_P strP, const char *value) {
-    size_t n = print_name(out);
-    char buf[16];
-    memset(buf, 0, 16);
-    strcpy_P(buf, strP);
-    n += print(out, buf);
-    n += print(out, ':');
+size_t Named::say_strP(PGM_P strP, int value) {
+    String str = String(value, DEC);
+    return say_strP(strP, (char *)str.c_str());
+}
+
+size_t Named::say_strP(PGM_P strP, char *value) {
+    size_t n = print_ident(out, name);
     n += print(out, ' ');
-    strcpy(buf, value);
-    n += print(out, buf);
-    return n += out->println();
+    n += PrintUtils::print_nameP_value(out, strP, value);
+    return n;
 }
 
 size_t AppModule::printDiag(Print *p) { return p->println(F("<empty>")); };